asiudhajsd errors

+++ backend/src/entity/run_report.rs
@@ -1,9 +1,9 @@
 use sea_orm::entity::prelude::*;
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
-pub type EclipseLevel = Option<u8>;
+pub type EclipseLevel = Option<i64>;
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, DeriveActiveEnum, EnumIter)]
+#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, DeriveActiveEnum, EnumIter)]
 #[sea_orm(
     rs_type = "String",
     db_type = "String(StringLen::None)",
@@ -16,7 +16,7 @@ pub enum Ending {
     Obliteration,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, Deserialize, DeriveActiveEnum, EnumIter)]
+#[derive(Clone, Debug, PartialEq, Eq, Deserialize, Serialize, DeriveActiveEnum, EnumIter)]
 #[sea_orm(
     rs_type = "String",
     db_type = "String(StringLen::None)",
@@ -29,7 +29,7 @@ pub enum Difficulty {
     Eclipse,
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel, Deserialize)]
+#[derive(Clone, Debug, PartialEq, Eq, DeriveEntityModel, Deserialize, Serialize)]
 #[sea_orm(table_name = "run_reports")]
 pub struct Model {
     #[serde(default)]
@@ -43,52 +43,52 @@ pub struct Model {
     pub start_time: DateTime,
     pub ending: Ending,
     pub difficulty: Difficulty,
-    pub eclipse_level: Option<u8>, // cant be EclipseLevel for sea orm reasons 🐸🚀
-    pub time_alive_seconds: u64,
-    pub stages_completed: u64,
+    pub eclipse_level: Option<i64>, // cant be EclipseLevel for sea orm reasons 🐸🚀
+    pub time_alive_seconds: i64,
+    pub stages_completed: i64,
     // pub run_time_seconds: u32,
 
     // items
-    pub items_collected: u64,
+    pub items_collected: i64,
     // pub items_scrapped: u16,
     // pub chests_left_behind: u16,
 
     // drones
-    pub drones_purchased: u64,
-    pub turrets_purchased: u64,
+    pub drones_purchased: i64,
+    pub turrets_purchased: i64,
     // pub drones_scrapped: u16,
     // pub drone_deaths: u32,
     // pub drones_left_behind: u16,
 
     // combat
-    pub kills: u64,
-    pub elite_kills: u64,
-    pub minion_kills: u64,
-    pub deaths: u64,
+    pub kills: i64,
+    pub elite_kills: i64,
+    pub minion_kills: i64,
+    pub deaths: i64,
 
     // damage
-    pub damage_dealt: u64,
-    pub minion_damage_dealt: u64,
-    // pub minion_damage_taken: u64,
-    pub damage_taken: u64,
-    // pub damage_dealt_over_time: u64,
-    pub highest_damage_dealt: u64,
+    pub damage_dealt: i64,
+    pub minion_damage_dealt: i64,
+    // pub minion_damage_taken: i64,
+    pub damage_taken: i64,
+    // pub damage_dealt_over_time: i64,
+    pub highest_damage_dealt: i64,
     // pub damage_blocked: u32,
 
     // healing
-    pub healing_recieved: u64,
-    // pub minion_healing_recieved: u64,
+    pub healing_recieved: i64,
+    // pub minion_healing_recieved: i64,
 
     // progression
-    pub highest_level: u64,
-    pub gold_collected: u64,
-    pub gold_spent: u64,
-    pub lunar_coins_spent: u64,
-    pub purchases: u64,
-    pub blood_purchases: u64,
+    pub highest_level: i64,
+    pub gold_collected: i64,
+    pub gold_spent: i64,
+    pub lunar_coins_spent: i64,
+    pub purchases: i64,
+    pub blood_purchases: i64,
 
     // movement
-    pub distance_traveled_metres: u64,
+    pub distance_traveled_metres: i64,
     // pub jumps: u32,
 
     // abilities
